PR Fix: Consolidated script code into converter.py
